requirements.txt
streamlit
langchain
langchain-openai
langchain-community
openai
chromadb
python-dotenv
beautifulsoup4